% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{add_contours}
\alias{add_contours}
\title{adds contours}
\usage{
add_contours(params, dimns, g)
}
\arguments{
\item{dimns}{grid layout (where each entry is a pixel) to build the picture}

\item{g}{a ggplot2 object}

\item{mcmcpath}{path to the mcmc output files}
}
\value{
ggplot2 object
}
