% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{add_map}
\alias{add_map}
\title{add_map}
\usage{
add_map(dimns, params, g)
}
\arguments{
\item{dimns}{the grid layout (nrow*ncol=npixels) to build the picture}
}
\value{
ggplot2 pobject
}
\description{
add_map
}
