% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_files_at_path}
\alias{check_files_at_path}
\title{checks if files exist}
\usage{
check_files_at_path(files, paths = ".", strict = F)
}
\arguments{
\item{files}{a list of files to check if they exist}
}
\description{
checks if files exist
}
