% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_contour_vals}
\alias{compute_contour_vals}
\title{compute contour vals for each pixel}
\usage{
compute_contour_vals(marks, now_rates, now_seeds)
}
\arguments{
\item{marks}{matrix of pixels}

\item{now_rates}{the rate of each voronoi tile}

\item{now_seeds}{the locations of each voronoi tile}
}
\value{
values vectoring storing the value for each pixel
}
\description{
compute one contour, by filling in each of the pixels/marks
}
