% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_scaling}
\alias{compute_scaling}
\title{computes scaling factors so results are interpretable in physical quanitites.}
\usage{
compute_scaling(mcmcpath)
}
\arguments{
\item{mcmcpath}{path to mcmc files}
}
\description{
computes scaling factors so results are interpretable in physical quanitites.
}
