% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{make_base}
\alias{make_base}
\title{make_base}
\usage{
make_base(dimns, params, g)
}
\arguments{
\item{dimns}{the grid layout (nrow*ncol=npixels) to build the picture}
}
\value{
ggplot2 pobject
}
\description{
make_base
}
