% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_fit_data}
\alias{plot_fit_data}
\title{plots the diagnostic files to assess whether MAPS fits the data "well"}
\usage{
plot_fit_data(mcmcpath, outpath, longlat)
}
\arguments{
\item{params}{list of parameter options}
}
\description{
plots the diagnostic files to assess whether MAPS fits the data "well"
}
