% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{plot_voronoi_samples}
\alias{plot_voronoi_samples}
\title{plot_voronoi_samples}
\usage{
plot_voronoi_samples(longlat, mcmcpath, outpath, is.mrates = TRUE,
  nsamples = 10)
}
\arguments{
\item{longlat}{Add description of longlat here.}

\item{mcmcpath}{Add description of mcmcpath here.}

\item{outpath}{Add description of outpath here.}

\item{is.mrates}{Add description of is.mrates here.}
}
\description{
plot_voronoi_samples
}
