% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_graph}
\alias{read_graph}
\title{reads the underlying graph}
\usage{
read_graph(path, longlat)
}
\arguments{
\item{path}{path to mcmc output files}

\item{longlat}{a binary variable, TRUE if long appears before lat}
}
\description{
reads the underlying graph
}
