% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{plot_maps}
\alias{plot_maps}
\title{plot_maps maining plotting function}
\usage{
plot_maps(add.pts = TRUE, add.graph = TRUE, add.countries = FALSE,
  plot.median = TRUE, longlat, mcmcpath, outpath, width = 10, height = 6,
  plot.difference = FALSE, set.range = TRUE)
}
\arguments{
\item{add.pts}{mark demes that are sampled (boolean)}

\item{add.graph}{overlay the underlying graph (boolean)}

\item{add.countries}{add country borders (boolean)}

\item{longlat}{order of longitude/lattitude in the .coords and .outer file (boolean)}

\item{mcmcpath}{path to mcmc output files}

\item{outpath}{files will be written to this path}

\item{width}{width of the main MAPS plot}

\item{height}{height of the main MAPS plot}

\item{plot.difference}{set TRUE if user ran MAPS with the olderpath parameter set, this way
MAPS only estimates the difference (boolean)}

\item{set.range}{TRUE if minimum variability is set as +- 1 (in log10 scale) from the mean (boolean)}

\item{plot.mean}{plots the median if TRUE and mean if FALSE (boolean)}
}
\description{
plot_maps maining plotting function
}
