% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trace}
\alias{plot_trace}
\title{plots the log-likelihood and log-posterior as a function of MCMC iteration (thinned)}
\usage{
plot_trace(mcmcpath, outpath)
}
\arguments{
\item{mcmcpath}{path to output mcmc files}

\item{outpath}{where to save the file}
}
\description{
plots the log-likelihood and log-posterior as a function of MCMC iteration (thinned)
}
