% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{plot_voronoi_samples}
\alias{plot_voronoi_samples}
\title{plot_voronoi_samples}
\usage{
plot_voronoi_samples(longlat, mcmcpath, outpath, is.mrates = TRUE,
  nsamples = 10, add.countries = TRUE, width = 5, height = 5)
}
\arguments{
\item{longlat}{order of longitude/lattitude in the .coords and .outer file (boolean)}

\item{mcmcpath}{path to mcmc output files}

\item{outpath}{files will be written to this path}

\item{is.mrates}{plot mvoronoi (TRUE) or qvoronoi (FALSE)}

\item{nsamples}{number of samples to plot from posterior}

\item{add.countries}{add country borders (boolean)}

\item{width}{width of the main MAPS plot}

\item{height}{height of the main MAPS plot}
}
\description{
plot_voronoi_samples
}
